rename TagForm.js
