+++ src/components/views/ApplicationViews.js
@@ -4,7 +4,7 @@ import { Login } from "../auth/Login"
 import { Register } from "../auth/Register"
 import { AuthorizedRoute } from "../auth/AuthorizedRoute"
 import { NavBar } from "../nav/NavBar"
-import { CreateTagForm } from "./tags/CreateTagForm.js"
+import { CreateTagForm } from "./tags/TagForm.js"
 import { TagList } from "./tags/TagList.js"
 import { PostDetails } from "./posts/PostDetails"
 import { UserPostList } from "./posts/UserPosts.js"
